Use single docker-compose.yml to run all of the e2e tests instead of mixing gitlab services and e2e compose services for ease of use and simplicity. Build a dedicated e2e test image so the normal web-frontend build is not slowed down by image building. Fix small issues and errors in e2e tests setup.

+++ e2e-tests/tests/auth/login.spec.ts
@@ -10,7 +10,12 @@ test.beforeEach(async () => {
 })
 
 test.afterEach(async () => {
-  await deleteUser(user)
+  // We only want to bother cleaning up in a devs local env or when pointed at a real
+  // server. If in CI then the first user will be the first admin and this will fail.
+  // Secondly in CI we are going to delete the database anyway so no need to clean-up.
+  if(!process.env.CI){
+    await deleteUser(user)
+  }
 })
 
 test('User can log in with email/password', async ({ page }) => {
